Fixed some import issues

+++ shell/import_from_access.php
@@ -17,6 +17,11 @@ $tables = array(
     //Foms
     'schede_a',
     'schede_b',
+    'sched_b1',
+    'sched_b2',
+    'sched_b3',
+    'sched_b4',
+    'schede_x',
     'schede_c',
     'schede_d',
     'schede_e',
@@ -24,11 +29,6 @@ $tables = array(
     'schede_g',
     'schede_g1',
     'schede_n',
-    'schede_x',
-    'sched_b1',
-    'sched_b2',
-    'sched_b3',
-    'sched_b4',
     'sched_c1',
     'sched_c2',
     'sched_d1',
@@ -77,6 +77,12 @@ $tables = array(
 if ($argc < 1) {
     echo 'MDB path is required';
     exit;
+} else if ($argc > 2) {
+    foreach ($argv as $key=>$value) {
+        if ($key == 0) continue;
+        else $argv[1].='\\ '.$value;
+        
+    }
 }
 $config = <<<EOL
 [pgsql]     
@@ -93,7 +99,6 @@ copy_every      = 100
 [tmpl]
 template     = True
 format       = csv
-datestyle    = dmy
 field_sep    = ,
 trailing_sep = True
 
@@ -103,23 +108,66 @@ if(!is_dir(__DIR__.DIRECTORY_SEPARATOR.'output'))
 
 foreach ($tables as $table) {
     $filename = __DIR__.DIRECTORY_SEPARATOR.'output'.DIRECTORY_SEPARATOR.$table.'.txt';
+    $filename_t = __DIR__.DIRECTORY_SEPARATOR.'output'.DIRECTORY_SEPARATOR.$table.'_t.txt';
     $logname = __DIR__.DIRECTORY_SEPARATOR.'output'.DIRECTORY_SEPARATOR.$table.'.log';
     $dataname = __DIR__.DIRECTORY_SEPARATOR.'output'.DIRECTORY_SEPARATOR.$table.'.data';
     if (is_file($filename)) unlink($filename);
+    if (is_file($filename_t)) unlink($filename_t);
     if (is_file($logname)) unlink($logname);
     if (is_file($dataname)) unlink($dataname);
-    exec('mdb-export -H '.$argv[1].' '.$table.' > '.$filename);
-    $config .= <<<EOL
+    exec('mdb-export -D "%Y-%m-%d %H:%M:%S" '.$argv[1].' '.$table.' > '.$filename);
+    
+    $file_input = fopen ($filename,'r');
+    $file_output = fopen ($filename_t,'w');
+    $p = ftell($file_input);
+    $columns_str = '';
+    if (is_resource($file_input)) {
+        while ($row = fgets($file_input)) {
+            if ($p == 0) {
+                $db_cols = $db->fetchCol('SELECT column_name FROM information_schema.columns WHERE table_name =\''.$table.'\'');
+                $db_cols = array_reverse($db_cols);
+                $columns = str_replace("\n",'',strtolower($row));
+                $columns = explode(',',$columns);
+                foreach ($db_cols as $db_col) {
+                    $key = array_search($db_col, $columns);
+                    if ($key === false) continue;
+                    if ($columns_str != '')
+                        $columns_str .= ',';
+                    $columns_str .= $db_col.':'.($key+1);        
+                }
+            }
+            else {
+                preg_match_all('/[0-9]\.[0-9]*e\+[0-9]{2}/', $row,$e_matches);
+                if (key_exists(0, $e_matches)) {
+                    foreach($e_matches[0] as $e_match) {
+                        $row = str_replace($e_match, floatval($e_match), $row);
+                    }
+                }
+                fputs($file_output,$row);
+            }
+            $p = ftell($file_input);
+        }
+        fclose($file_input);
+        fclose($file_output);
+    }
+file_put_contents('pgloader.conf', $config. <<<EOL
 
 [{$table}]
 use_template    = tmpl
-table           = '{$table}'
-filename = {$filename}
+table           = {$table}
+filename = {$filename_t}
+columns = {$columns_str}
 reject_log = output/{$table}.log
 reject_data = output/{$table}.data
 
-EOL;
+EOL
+);
+exec('pgloader '.$table);
+if (is_file('pgloader.conf')) unlink('pgloader.conf');
+if (is_file($filename)) unlink($filename);
+if (is_file($filename_t)) unlink($filename_t);
+$error = file_exists($logname) && filesize($logname) > 0;
+if (is_file($logname) && !$error) unlink($logname);
+if (is_file($dataname) && !$error) unlink($dataname);
 }
-file_put_contents('pgloader.conf', $config);
-exec('pgloader');
-unlink('pgloader.conf');
+
